requirements.txt
Flask
requests>=2.25.1
boto3>=1.18.0
gunicorn
python-dotenv

